test: move crashReporter specs to the main process
